08/01/2022 Fin sección 12

+++ 07-journal/src/modules/daybook/views/EntryView.vue
@@ -7,17 +7,25 @@
 				<span class="mx-2 fw-light">{{ yearDay }}</span>
 			</div>
 			<div>
-				<button v-if="entry.id"
-								class="btn btn-danger mx-2"
-								@click="onDeleteEntry">Borrar <em class="fa fa-trash-alt"></em></button>
-				<button class="btn btn-primary mx-2">Subir foto <em class="fa fa-upload"></em></button>
+				
+				<input v-show="false" ref="imageSelector" accept="image/png, image/jpeg" type="file" @change="onSelectedImage">
+				
+				<button v-if="entry.id" class="btn btn-danger mx-2" @click="onDeleteEntry">
+					Borrar
+					<em class="fa fa-trash-alt"></em>
+				</button>
+				<button class="btn btn-primary mx-2" @click="onSelectImage">
+					Subir foto
+					<em class="fa fa-upload"></em>
+				</button>
 			</div>
 		</div>
 		<hr>
 		<div class="d-flex flex-column px-3 h-75">
 			<textarea v-model="entry.text" placeholder="Que sucedió hoy?"></textarea>
 		</div>
-		<img alt="entry-picture" class="img-thumbnail" src="http://qnimate.com/wp-content/uploads/2014/03/images2.jpg">
+		<img v-if="entry.picture && !localImg" :src="entry.picture" alt="entry-picture" class="img-thumbnail">
+		<img v-if="localImg" :src="localImg" alt="entry-picture" class="img-thumbnail">
 	</template>
 	<Fab icon="fa-save" @on:click="saveEntry" />
 </template>
@@ -26,6 +34,8 @@
 import { defineAsyncComponent }   from 'vue';
 import { mapActions, mapGetters } from "vuex";
 import getDayMonthYear            from "@/modules/daybook/helpers/getDayMonthYear.ts";
+import Swal                       from "sweetalert2";
+import uploadImage                from "@/modules/daybook/helpers/uploadImage";
 
 export default {
 	name      : 'EntryView',
@@ -40,9 +50,11 @@ export default {
 	},
 	data() {
 		return {
-			entry: {
+			entry   : {
 				text: ''
-			}
+			},
+			localImg: '',
+			file    : null
 		}
 	},
 	computed: {
@@ -87,16 +99,58 @@ export default {
 			}
 		},
 		async saveEntry() {
+			new Swal( {
+				title            : 'Espere por favor',
+				allowOutsideClick: false
+			} );
+			Swal.showLoading();
+			
+			this.entry.picture = await uploadImage( this.file );
+			
 			if ( this.entry.id ) {
 				await this.updateEntry( this.entry );
 			} else {
 				const id = await this.createEntry( this.entry );
 				await this.$router.push( { name: 'entry', params: { id } } );
 			}
+			
+			this.file = null;
+			Swal.fire( 'Guardando', 'Entrada registrada con exito', 'success' );
 		},
 		async onDeleteEntry() {
-			await this.deleteEntry( this.entry );
-			await this.$router.push( { name: 'no-entry' } );
+			const { isConfirmed } = await Swal.fire( {
+				title            : 'Esta seguro?',
+				text             : 'Una vez borrado, no se puede recuperar',
+				showDenyButton   : true,
+				confirmButtonText: 'Si, estoy seguro'
+			} );
+			
+			if ( isConfirmed ) {
+				new Swal( {
+					title            : 'Espere por favor',
+					allowOutsideClick: false
+				} );
+				
+				await this.deleteEntry( this.entry );
+				await this.$router.push( { name: 'no-entry' } );
+				
+				Swal.fire( 'Eliminado', '', 'success' );
+			}
+		},
+		onSelectedImage( event ) {
+			const file = event.target.files[ 0 ];
+			if ( !file ) {
+				return;
+			}
+			
+			const fr  = new FileReader();
+			fr.onload = () => this.localImg = fr.result;
+			fr.readAsDataURL( file );
+			this.file = file;
+		},
+		onSelectImage() {
+			const { imageSelector } = this.$refs;
+			imageSelector.click();
 		}
 	},
 	created() {
